implement ReadString properly
turns out the naive assumption that BinaryReader.ReadString would just do
the right thing was false, and it simply worked by chance before (since
apparently, none of the test files by now had a string with a length
exceeding 7 bits).

explicitly reading the length and then reading characters one-by-one seems
to work a lot better, considering that BinaryReader expects the string
length to be encoded as 7-bit integer; rather than being 255 which is the
marker for "long length" inside a metafile.