implement handling of long command headers
this is needed for other implementations and tries to hide this encoding
detail by reading the whole element into a buffer, which is then presented
to the according reader as if it were a short command without partitioning.

it also lets them treat EOF as a lazy exit condition, rather than having to
count bytes read for commands that contain variable data lists (such as
FONT LIST).