TILE/gui: try to support TIFF-like compressed TILEs
the majority of the remaining compression types are supported by TIFF
(primarily CCITT Group 3 T4 and Group 4 T6), so lets try to generate a valid
TIFF and use the built-in TIFF decoder.

the TIFF Generator kind-of works, but is very quirky for the originally
intended target of CCITT. Most CCITT images end up single-color for whatever
reason; but at least they don't throw.

we can also use this for uncompressed (Bitmap) types, which seem fine mostly
but sometimes have color issues or shifted image lines.

Right now, it seems better than nothing at all, so we'll keep it until a
better solution (or fix for it) comes around.