Text Encoding: fix reading of empty strings that are followed by hard separators
previously, we would back up once and in the following loop determine we're
at the end, but this caused empty strings inside of SDRs to be read as EOF
instead of just the end of a token.
by skipping out when we know this is an empty string, we can avoid backing
up and fix it in the process.