Binary: fix writing of commands and strings that exceed the short-form length
commands may exceed 0x1F, strings may exceed 0xFF which causes them to
require the use of long-form representation.
this is the short-form length set to the maximum, followed by 15 bit length
with 1 bit continuation indicator and a block of the specified length;
repeating until everything is written to file (in 32k chunks each).